# [libc++] Optimize ranges::copy{, _n} for vector<bool>::iterator