--- a/libcxx/test/std/algorithms/alg.modifying.operations/alg.copy/copy.pass.cpp
+++ b/libcxx/test/std/algorithms/alg.modifying.operations/alg.copy/copy.pass.cpp
@@ -60,27 +60,25 @@ struct TestInIters {
   }
 };
 
-template <std::size_t N>
-struct TestBitIter {
-  std::vector<bool> in;
-  TEST_CONSTEXPR_CXX20 TestBitIter() : in(N, false) {
-    for (std::size_t i = 0; i < N; i += 2)
-      in[i] = true;
+TEST_CONSTEXPR_CXX20 bool test_vector_bool(std::size_t N) {
+  std::vector<bool> in(N, false);
+  for (std::size_t i = 0; i < N; i += 2)
+    in[i] = true;
+
+  { // Test copy with aligned bytes
+    std::vector<bool> out(N);
+    std::copy(in.begin(), in.end(), out.begin());
+    assert(in == out);
   }
-  TEST_CONSTEXPR_CXX20 void operator()() {
-    { // Test copy with aligned bytes
-      std::vector<bool> out(N);
-      std::copy(in.begin(), in.end(), out.begin());
-      assert(in == out);
-    }
-    { // Test copy with unaligned bytes
-      std::vector<bool> out(N + 8);
-      std::copy(in.begin(), in.end(), out.begin() + 4);
-      for (std::size_t i = 0; i < N; ++i)
-        assert(out[i + 4] == in[i]);
-    }
+  { // Test copy with unaligned bytes
+    std::vector<bool> out(N + 8);
+    std::copy(in.begin(), in.end(), out.begin() + 4);
+    for (std::size_t i = 0; i < N; ++i)
+      assert(out[i + 4] == in[i]);
   }
-};
+
+  return true;
+}
 
 TEST_CONSTEXPR_CXX20 bool test() {
   types::for_each(types::cpp17_input_iterator_list<int*>(), TestInIters());
@@ -102,11 +100,11 @@ TEST_CONSTEXPR_CXX20 bool test() {
   }
 
   { // Test vector<bool>::iterator optimization
-    TestBitIter<8>()();
-    TestBitIter<16>()();
-    TestBitIter<32>()();
-    TestBitIter<64>()();
-    TestBitIter<1024>()();
+    assert(test_vector_bool(8));
+    assert(test_vector_bool(16));
+    assert(test_vector_bool(32));
+    assert(test_vector_bool(64));
+    assert(test_vector_bool(256));
   }
 
   return true;